+++ src/main/java/wooteco/subway/domain/pathfinder/PathFinder.java
@@ -1,6 +1,8 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.pathfinder;
 
 import java.util.List;
+import wooteco.subway.domain.Section;
+import wooteco.subway.domain.Station;
 
 public interface PathFinder {
