+++ src/main/java/wooteco/subway/service/PathService.java
@@ -3,8 +3,8 @@ package wooteco.subway.service;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
-import wooteco.subway.domain.PathFinder;
-import wooteco.subway.domain.PathFinderFactory;
+import wooteco.subway.domain.pathfinder.PathFinder;
+import wooteco.subway.domain.pathfinder.PathFinderFactory;
 import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Station;
 import wooteco.subway.domain.policy.DiscountPolicy;
